Issue #83 fix multithreading DaySimController bug
In RunChoiceModels in Engine.cs the variable index is used to say if a household is within the range sent by DaySimController. There are two problems with this code both stemming from the loop being a Parallel.For and therefore multithreaded. One is that the order the batches will execute in is random. The other is that the increment of index is not threadsafe.